packages/kcastle/pyproject.toml
[project]
name = "kcastle"
version = "0.1.0"
description = "Agent application with multi-endpoint support for the k agent framework"
readme = "README.md"
authors = [
    { name = "Xiangzhuang Shen", email = "datahonor@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "kagent",
    "pyyaml>=6.0",
]

[project.scripts]
k = "kcastle.cli:main"

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0", "telegramify-markdown>=0.5.4"]
all = ["python-telegram-bot>=21.0", "telegramify-markdown>=0.5.4"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
kagent = { workspace = true }

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
include = [
    "src/**/*.py",
    "tests/**/*.py",
]
